+++ R/layout.R
@@ -22,6 +22,10 @@ layout.matrix <- function(p, ..., data = NULL) {
 layout.plotly <- function(p, ..., data = NULL) {
   p <- add_data(p, data)
   attrs <- list(...)
+  if (!is.null(attrs[["height"]]) || !is.null(attrs[["width"]])) {
+    warning("Specifying width/height in layout() is now deprecated.\n", 
+            "Please specify in ggplotly() or plot_ly()", call. = FALSE)
+  }
   # similar to add_trace()
   p$x$layoutAttrs <- c(
     p$x$layoutAttrs %||% list(), 
@@ -57,13 +61,45 @@ rangeslider <- function(p, ...) {
 #' @param p a plotly object
 #' @param ... these arguments are documented at 
 #' \url{https://github.com/plotly/plotly.js/blob/master/src/plot_api/plot_config.js}
+#' @param collaborate include the collaborate mode bar button (unique to the R pkg)?
+#' @param cloud include the send data to cloud button?
 #' @author Carson Sievert
 #' @export
-#' @examples \dontrun{
-#' config(plot_ly(), displaylogo = FALSE, modeBarButtonsToRemove = list('sendDataToCloud'))
-#' }
+#' @examples
+#' 
+#' config(plot_ly(), displaylogo = FALSE, collaborate = FALSE)
+#' 
 
-config <- function(p, ...) {
+config <- function(p, ..., collaborate = TRUE, cloud = FALSE) {
+  
   p$x$config <- modify_list(p$x$config, list(...))
+  
+  nms <- sapply(p$x$config[["modeBarButtonsToAdd"]], "[[", "name")
+  hasCollab <- sharingButton()[["name"]] %in% nms
+  
+  if (collaborate && !hasCollab) {
+    nAdd <- length(p$x$config[["modeBarButtonsToAdd"]])
+    p$x$config[["modeBarButtonsToAdd"]][[nAdd + 1]] <- sharingButton()
+  }
+  if (!collaborate) {
+    p$x$config[["modeBarButtonsToAdd"]][nms %in% sharingButton()[["name"]]] <- NULL
+  }
+
+  hasCloud <- !'sendDataToCloud' %in% p$x$config[["modeBarButtonsToRemove"]]
+  if (!cloud) {
+    p$x$config[["modeBarButtonsToRemove"]] <- c(
+      p$x$config[["modeBarButtonsToRemove"]], 'sendDataToCloud'
+    )
+  }
+  if (cloud) {
+    b <- p$x$config[["modeBarButtonsToRemove"]]
+    p$x$config[["modeBarButtonsToRemove"]] <- b[!b %in% 'sendDataToCloud']
+  }
+  
+  # ensure array
+  if (length(p$x$config[["modeBarButtonsToRemove"]]) == 1) {
+    p$x$config[["modeBarButtonsToRemove"]] <- list(p$x$config[["modeBarButtonsToRemove"]])
+  }
+
   p
 }